extracted get_moon() to lib/view/moon.py
